feat: 프로필별 출력
