+++ 안예지/0213/quick-start/src/App.js
@@ -1,8 +1,10 @@
 import './App.css';
+import Users from './components/ProfileList';
 
 function App() {
   return (
     <div className="App">
+      <Users />
 
 
     </div>
