+++ 안예지/0213/quick-start/src/components/Profile.js
@@ -1,15 +1,11 @@
-const userList = [{
-    id: 1,
-    name: '류진숙',
-    imageUrl: 'https://dummyimage.com/90/bdaaeb/fffce3&text=dahila',
-},
-{
-    id: 2,
-    name: '박혜진',
-    imageUrl: 'https://dummyimage.com/90/eba7b3/fffce3&text=hye',
-},
-{
-    id: 3,
-    name: '안예지',
-    imageUrl: 'https://dummyimage.com/90/b7ebab/fffce3&text=avocado',
-},]
+const Profile = ({ name, image }) => {
+    return (
+        <div className="user">
+            <div>이름 : {name}</div>
+            <img src={image} />
+        </div>
+
+    )
+}
+
+export default Profile
